requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
appnope==0.1.4
argparse-addons==0.12.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitstruct==8.19.0
bleach==6.2.0
blinker==1.8.2
cachetools==5.5.0
cantools==39.4.11
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colour==0.1.5
crccheck==1.3.0
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
digi==0.2.8
digi-xbee==1.5.0
diskcache==5.6.3
docopt==0.6.2
durationpy==0.9
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.0.3
frozenlist==1.5.0
google-auth==2.35.0
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
ipython==8.12.3
iso8601==2.1.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kopf==1.37.2
kubernetes==31.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.0.8
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.1.3
oauthlib==3.2.2
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyserial==3.5
python-box==7.2.0
python-can==4.4.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzed==1.3.0
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rpds-py==0.20.1
rsa==4.9
setuptools==75.3.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
tenacity==9.0.0
textparser==0.24.0
tinycss2==1.4.0
tkdial==0.0.7
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
wrapt==1.16.0
yarg==0.1.9
yarl==1.17.1
zipp==3.20.2
